Allow printing expressions based on feature set
As discussed in #3517, we need a way to maintain the ability to print
expressions in a backward-compatible way. This adds a FeatureSet
to PrintExpressionContents which defaults to FeatureSet::All in case
a feature set is not available.

This will be used for CallIndirect to decide whether to print the table
name arg.