[Wasm Exceptions] Fix RemoveUnusedNames on Try (#3583)
The delegate field of Try was not being scanned, so we could
remove a name that was used only by a delegate.

The bug was because visitTry overrides the generic visitor
visitExpression. So we need to call it manually. Sadly the code here
was pretty old (I probably wrote it back in 2015 or so) and it was
misleading, as it had unnecessary calls from the generic visitor
to visitBlock, visitLoop, which are not needed. This PR removes
them which is shorter and cleaner.

Also, we must handle the case of the delegate field being unset,
so check name.is().