Add C and JS API functions for accessing memory info (#3573)
This allows identifying whether a memory is present on the module,
getting its initial and, if present, maximum sizes, and whether it's
shared as well as its import module and base name if imported.

Will require a matching update to index.d.ts in binaryen.js for
type hinting when used via npm.