Properly use text format type names in printing (#3591)
This adds a TypeNames entry to modules, which can store names for types. So
far this PR uses that to store type names from text format. Future PRs will add
support for field names and for the binary format.

(Field names are added to wasm.h here to see if we agree on this direction.)

Most of the work here is threading a module through the various functions in
Print.cpp. This keeps the module optional, so that we can still print an
expression independently of a module, which has always been the case, and
which I think we should keep (but, if a module was mandatory perhaps this
would be a little simpler, and could be refactored into a form that depends on
that).

99% of this diff are test updates, since almost all our tests use the text
format, and many of them specify a type name but we used to ignore it.

This is a step towards a proper solution for #3589