Remove unneeded static_assert on number of Names in hashing (#3606)
The assertion is not really needed. Wasm64 will need changes to support
more than 2^32 names, in theory, but (1) wasm64 is just memory64 atm,
and (2) we'd need to add a general option for Index to be larger than 32
bits in general, so there is nothing specific to the hashing code here.