[Wasm Exceptions] Properly ensure unique Try labels after an inlining (#3632)
The old code here just referred to Block and Loop. Refactor it to use the
generic helper code that also handles Try.

Also add validation of Try names in the validator.

The testcase here would have $label appear twice before this fix. After
the fix there is $label0 for one of them.