[Wasm GC] Add test/spec/br_on_null.wast and validation fixes for it (#3623)
This adds ValidationBuilder which can allow sharing of builder code that also
validates, between the text and binary parsers. In general we share that code in
the validator, but the validator can only run once IR exists, and in some cases we
can't even emit valid IR structure at all.