[Wasm GC] Add Names section support for field names (#3589)
Adds support for GC struct fields in the binary format, implementing
WebAssembly/gc#193

No extra tests needed, see the .fromBinary output which shows this working.

This also has a minor fix in the s-parser, we should not always add a name
to the map of index=>name - only if it exists. Without that fix, the binary
emitter would write out null strings.