[Wasm GC] Properly handle "typeindex" in the binary format (#3663)
We handled them as S63 instead of U32. That should be fine, as all U32 values fit
in S63. But it is not strictly correct. The signed encoding may use an additional byte
which is unnecessary, and there is an actual correctness issue where a U32 may
be interpreted as a large negative S63 (because it sign extends a final bit that
happens to be 1).

May help #3656 but that testcase still does not pass even with this.