[Wasm GC] Optimize BrOn* of the wrong kind (#3724)
#3719 optimized the case where the kind is what we want, like br_on_func of
a function. This handles the opposite case, where we know the kind is wrong, and
so the break is not taken.

This seems equally useful for "polymorphic" code that does a bunch of checks
and routes to different code for each case, as after inlining or other optimizations
we may see which paths are taken and which are not.

Also refactor the checking code to a shared location as RefIs/As will also want to
use it.